+++ packages/ensnode-sdk/src/internal.ts
@@ -1,0 +1,10 @@
+/**
+ * Internal APIs
+ *
+ * This file can never be included in the NPM package for ENSNode SDK.
+ *
+ * The only way to import functionality from this file is to
+ * use `@ensnode/ensnode-sdk/internal` path. This path is available in any
+ * app/package in the monorepo which requires `@ensnode/ensnode-sdk` dependency.
+ */
+export { BlockRefSchema, ChainIdSchema, DatetimeSchema, UrlSchema } from "./shared/zod-schemas";
