+++ packages/ensnode-sdk/src/shared/index.ts
@@ -1,2 +1,8 @@
 export * from "./domain-types";
 export * from "./serialized-types";
+export {
+  deserializeBlockRef,
+  deserializeChainIdString,
+  deserializeDatetime,
+  deserializeUrl,
+} from "./deserialize";
